+++ changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## WIP
+  - Use newer version of `checkers`
 
 ## [0.12.4] - 2020-06-07
   - Big speedup (orbits testsuite about 9 times faster)